+++ WebApplicationAPI/Models/Certificate/Region.cs
@@ -9,7 +9,7 @@ namespace WebApplicationAPI.Models.Certificate
     public class Region
     {
         [Key]
-        public int RegionID { get; set; }
+        public long RegionID { get; set; }
         public string? RegionName { get; set; }
     }
 }
