+++ WebApplicationAPI/Service/Interfaces/IAreaService.cs
@@ -5,6 +5,6 @@ namespace WebApplicationAPI.Service.Interfaces
     public interface IAreaService
     {
         Task<List<AreaDTO>> GetAreas();
-        Task<AreaDTO> GetArea(int id);
+        Task<AreaDTO> GetArea(long id);
     }
 }
